server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
supabase==1.3.0
requests==2.31.0
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4
google-oauth2-tool==0.0.3
dlib==19.24.2
cmake==3.27.7